Add branch for the Freedesktop runtime 24.08
This removes the appstream stuff, which is not present in this version
of the Freedesktop SDK, and it also adds two patches to fix build
errors given by GCC 14.

A special thanks to Gentoo dev, Sam James, which backported these patches
from upstream a few months ago:

https://gitweb.gentoo.org/repo/gentoo.git/commit/?id=250d2f45f958ccb2a4e3c437d396b88f5d3bd59e

Please note that I have not included the other two patches on purpose,
because as of now they're not needed to make the build pass.